requirements.txt
fastapi
uvicorn
PyPDF2
pytesseract
Pillow
python-multipart